BE/requirements.txt
amqp==5.3.1
asgiref==3.10.0
billiard==4.2.2
black==25.9.0
boto3==1.40.63
botocore==1.40.63
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
flake8==7.3.0
idna==3.11
jmespath==1.0.1
kombu==5.5.4
mccabe==0.7.0
mypy_extensions==1.1.0
mysqlclient==2.2.7
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
pycodestyle==2.14.0
pyflakes==3.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.2.0
requests==2.32.5
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
